* Added the SERVER_RESOURCE table (didn't want resources being internalized, as 1.098 was. That was messy.)  * Completely removed zone indexes. As zone IDs were unique regardless of the server they were on, the whole system seems pointless (slow not having it indexed also).  * The level up table now uses a map, instead of a vector so that they're better & more safely indexed.  * Merged NPC in-type packet data with CNpc::GetNpcInfo(), as with CUser::GetUserInfo().

+++ Ebenezer/EVENT.cpp
@@ -52,7 +52,7 @@ BOOL EVENT::LoadEvent(int zone)
 
 	if( !pFile.Open( filename, CFile::modeRead) ) return TRUE;
 
-	length = pFile.GetLength();
+	length = (DWORD)pFile.GetLength();
 	
 	CArchive in(&pFile, CArchive::load);
